Backend/Media_Recommendation/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
Flask==3.1.2
flask-cors==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
RapidFuzz==3.14.1
redis==6.4.0
requests==2.32.5
spotipy==2.25.1
urllib3==2.5.0
Werkzeug==3.1.3
